fixed narration and organized classes into packages. also fixed a mute/unmute bug
