+++ app/src/main/java/com/example/myfirstapp/Activity/Activity_Menu_Backstory1.java
@@ -1,10 +1,4 @@
-/*
- ***************************************************
- * see Activity_Menu_Backstory1 for details
- ***************************************************
- */
-
-package com.example.myfirstapp;
+package com.example.myfirstapp.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -14,15 +8,24 @@ import android.os.Bundle;
 import android.view.View;
 import android.view.WindowManager;
 
-public class Activity_Menu_Backstory3 extends AppCompatActivity {
+import com.example.myfirstapp.Audio.Audio_Master_Control;
+import com.example.myfirstapp.R;
+
+public class Activity_Menu_Backstory1 extends AppCompatActivity {
+    /**
+     * to play audio
+     */
     private MediaPlayer myNarration;
     private MediaPlayer myBGM;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_backstory3);
+        //set screen view to this
+        setContentView(R.layout.activity_backstory1);
+        //don't show notification bar
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
 
+        //initialize audio
         myNarration = MediaPlayer.create(this, R.raw.backstory1);
         myBGM = MediaPlayer.create(this, R.raw.bgm_modes_loop);
         if(Audio_Master_Control.myMuted)
@@ -37,10 +40,11 @@ public class Activity_Menu_Backstory3 extends AppCompatActivity {
         }
         myNarration.start();
         myBGM.start();
-
-        findViewById(R.id.previous3).setOnClickListener(new View.OnClickListener() {
+        //onclick listener for skip (go straight to play)
+        findViewById(R.id.skip).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                //release the resources associated with this audio player
                 if(myNarration !=null) {
                     if(myNarration.isPlaying())
                         myNarration.stop();
@@ -55,11 +59,37 @@ public class Activity_Menu_Backstory3 extends AppCompatActivity {
                     myBGM.release();
                     myBGM =null;
                 }
-                startActivity(new Intent(Activity_Menu_Backstory3.this, Activity_Menu_Backstory2.class));
+                //show main game activity
+                startActivity(new Intent(Activity_Menu_Backstory1.this, Activity_Game.class));
             }
         });
 
-        findViewById(R.id.playNow).setOnClickListener(new View.OnClickListener() {
+        //next backstory
+        findViewById(R.id.continue1).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                //release the resources associated with this audio player
+                if(myNarration !=null) {
+                    if(myNarration.isPlaying())
+                        myNarration.stop();
+                    myNarration.reset();
+                    myNarration.release();
+                    myNarration =null;
+                }
+                if(myBGM !=null) {
+                    if(myBGM.isPlaying())
+                        myBGM.stop();
+                    myBGM.reset();
+                    myBGM.release();
+                    myBGM =null;
+                }
+                //show main game activity
+                startActivity(new Intent(Activity_Menu_Backstory1.this, Activity_Menu_Backstory2.class));
+            }
+        });
+
+        //back to main
+        findViewById(R.id.previous1).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 if(myNarration !=null) {
@@ -76,7 +106,7 @@ public class Activity_Menu_Backstory3 extends AppCompatActivity {
                     myBGM.release();
                     myBGM =null;
                 }
-                startActivity(new Intent(Activity_Menu_Backstory3.this, Activity_Game.class));
+                startActivity(new Intent(Activity_Menu_Backstory1.this, Activity_Menu_Modes.class));
             }
         });
     }
