+++ app/src/main/java/com/example/myfirstapp/Activity/Activity_Menu_Backstory2.java
@@ -1,4 +1,9 @@
-package com.example.myfirstapp;
+/*
+ ***************************************************
+ * see Activity_Menu_Backstory1 for details
+ ***************************************************
+ */
+package com.example.myfirstapp.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -8,22 +13,19 @@ import android.os.Bundle;
 import android.view.View;
 import android.view.WindowManager;
 
-public class Activity_Menu_Backstory1 extends AppCompatActivity {
-    /**
-     * to play audio
-     */
+import com.example.myfirstapp.Audio.Audio_Master_Control;
+import com.example.myfirstapp.R;
+
+public class Activity_Menu_Backstory2 extends AppCompatActivity {
     private MediaPlayer myNarration;
     private MediaPlayer myBGM;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        //set screen view to this
-        setContentView(R.layout.activity_backstory1);
-        //don't show notification bar
+        setContentView(R.layout.activity_backstory2);
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
 
-        //initialize audio
-        myNarration = MediaPlayer.create(this, R.raw.backstory1);
+        myNarration = MediaPlayer.create(this, R.raw.backstory2);
         myBGM = MediaPlayer.create(this, R.raw.bgm_modes_loop);
         if(Audio_Master_Control.myMuted)
         {
@@ -37,11 +39,11 @@ public class Activity_Menu_Backstory1 extends AppCompatActivity {
         }
         myNarration.start();
         myBGM.start();
-        //onclick listener for skip (go straight to play)
+
+
         findViewById(R.id.skip).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                //release the resources associated with this audio player
                 if(myNarration !=null) {
                     if(myNarration.isPlaying())
                         myNarration.stop();
@@ -56,37 +58,11 @@ public class Activity_Menu_Backstory1 extends AppCompatActivity {
                     myBGM.release();
                     myBGM =null;
                 }
-                //show main game activity
-                startActivity(new Intent(Activity_Menu_Backstory1.this, Activity_Game.class));
+                startActivity(new Intent(Activity_Menu_Backstory2.this, Activity_Game.class));
             }
         });
 
-        //next backstory
-        findViewById(R.id.continue1).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //release the resources associated with this audio player
-                if(myNarration !=null) {
-                    if(myNarration.isPlaying())
-                        myNarration.stop();
-                    myNarration.reset();
-                    myNarration.release();
-                    myNarration =null;
-                }
-                if(myBGM !=null) {
-                    if(myBGM.isPlaying())
-                        myBGM.stop();
-                    myBGM.reset();
-                    myBGM.release();
-                    myBGM =null;
-                }
-                //show main game activity
-                startActivity(new Intent(Activity_Menu_Backstory1.this, Activity_Menu_Backstory2.class));
-            }
-        });
-
-        //back to main
-        findViewById(R.id.previous1).setOnClickListener(new View.OnClickListener() {
+        findViewById(R.id.previous2).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 if(myNarration !=null) {
@@ -103,7 +79,28 @@ public class Activity_Menu_Backstory1 extends AppCompatActivity {
                     myBGM.release();
                     myBGM =null;
                 }
-                startActivity(new Intent(Activity_Menu_Backstory1.this, Activity_Menu_Modes.class));
+                startActivity(new Intent(Activity_Menu_Backstory2.this, Activity_Menu_Backstory1.class));
+            }
+        });
+
+        findViewById(R.id.continue2).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(myNarration !=null) {
+                    if(myNarration.isPlaying())
+                        myNarration.stop();
+                    myNarration.reset();
+                    myNarration.release();
+                    myNarration =null;
+                }
+                if(myBGM !=null) {
+                    if(myBGM.isPlaying())
+                        myBGM.stop();
+                    myBGM.reset();
+                    myBGM.release();
+                    myBGM =null;
+                }
+                startActivity(new Intent(Activity_Menu_Backstory2.this, Activity_Menu_Backstory3.class));
             }
         });
     }
