+++ app/src/main/java/com/example/myfirstapp/Activity/Activity_Menu_Backstory3.java
@@ -3,7 +3,8 @@
  * see Activity_Menu_Backstory1 for details
  ***************************************************
  */
-package com.example.myfirstapp;
+
+package com.example.myfirstapp.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -13,16 +14,19 @@ import android.os.Bundle;
 import android.view.View;
 import android.view.WindowManager;
 
-public class Activity_Menu_Backstory2 extends AppCompatActivity {
+import com.example.myfirstapp.Audio.Audio_Master_Control;
+import com.example.myfirstapp.R;
+
+public class Activity_Menu_Backstory3 extends AppCompatActivity {
     private MediaPlayer myNarration;
     private MediaPlayer myBGM;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_backstory2);
+        setContentView(R.layout.activity_backstory3);
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
 
-        myNarration = MediaPlayer.create(this, R.raw.backstory1);
+        myNarration = MediaPlayer.create(this, R.raw.backstory3);
         myBGM = MediaPlayer.create(this, R.raw.bgm_modes_loop);
         if(Audio_Master_Control.myMuted)
         {
@@ -37,8 +41,7 @@ public class Activity_Menu_Backstory2 extends AppCompatActivity {
         myNarration.start();
         myBGM.start();
 
-
-        findViewById(R.id.skip).setOnClickListener(new View.OnClickListener() {
+        findViewById(R.id.previous3).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 if(myNarration !=null) {
@@ -55,11 +58,11 @@ public class Activity_Menu_Backstory2 extends AppCompatActivity {
                     myBGM.release();
                     myBGM =null;
                 }
-                startActivity(new Intent(Activity_Menu_Backstory2.this, Activity_Game.class));
+                startActivity(new Intent(Activity_Menu_Backstory3.this, Activity_Menu_Backstory2.class));
             }
         });
 
-        findViewById(R.id.previous2).setOnClickListener(new View.OnClickListener() {
+        findViewById(R.id.playNow).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 if(myNarration !=null) {
@@ -76,28 +79,7 @@ public class Activity_Menu_Backstory2 extends AppCompatActivity {
                     myBGM.release();
                     myBGM =null;
                 }
-                startActivity(new Intent(Activity_Menu_Backstory2.this, Activity_Menu_Backstory1.class));
-            }
-        });
-
-        findViewById(R.id.continue2).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if(myNarration !=null) {
-                    if(myNarration.isPlaying())
-                        myNarration.stop();
-                    myNarration.reset();
-                    myNarration.release();
-                    myNarration =null;
-                }
-                if(myBGM !=null) {
-                    if(myBGM.isPlaying())
-                        myBGM.stop();
-                    myBGM.reset();
-                    myBGM.release();
-                    myBGM =null;
-                }
-                startActivity(new Intent(Activity_Menu_Backstory2.this, Activity_Menu_Backstory3.class));
+                startActivity(new Intent(Activity_Menu_Backstory3.this, Activity_Game.class));
             }
         });
     }
