+++ app/src/main/java/com/example/myfirstapp/Game/Game_Heart.java
@@ -1,10 +1,12 @@
-package com.example.myfirstapp;
+package com.example.myfirstapp.Game;
 
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 
-import static com.example.myfirstapp.Game_Display.screenRatioX;
+import com.example.myfirstapp.R;
+
+import static com.example.myfirstapp.Game.Game_Display.screenRatioX;
 
 /**
  * drawing of life/heart
