+++ app/src/main/java/com/example/myfirstapp/Game/Game_MathProblems.java
@@ -1,4 +1,4 @@
-package com.example.myfirstapp;
+package com.example.myfirstapp.Game;
 
 import java.text.DecimalFormat;
 import java.util.HashSet;
