+++ app/src/main/res/layout/activity_backstory1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity_Menu_Backstory1">
+    tools:context=".Activity.Activity_Menu_Backstory1">
 
     <ImageView
         android:id="@+id/Step1"
