+++ app/src/main/res/layout/activity_backstory2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity_Menu_Backstory2">
+    tools:context=".Activity.Activity_Menu_Backstory2">
 
     <ImageView
         android:id="@+id/Step1"
