+++ app/src/main/res/layout/activity_credits.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="horizontal"
     android:background="@drawable/main_menu_background"
-    tools:context=".Activity_Menu_Credits">
+    tools:context=".Activity.Activity_Menu_Credits">
 
     <RelativeLayout
         android:layout_width="match_parent"
