+++ app/src/main/res/layout/activity_game_over.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity_Game_Over"
+    tools:context=".Activity.Activity_Game_Over"
     android:orientation="horizontal">
 
     <ImageView
