+++ app/src/main/res/layout/activity_modes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity_Menu_Modes">
+    tools:context=".Activity.Activity_Menu_Modes">
 
     <ImageView
         android:id="@+id/Step1"
